Combine flowmachine.core.utils and flowmachine.utils.utils into just flowmachine.utils
