+++ flowmachine/flowmachine/features/location/total_events.py
@@ -19,7 +19,7 @@ from ..utilities import EventsTablesUnion
 from ...core import Query
 from ...core.mixins import GeoDataMixin
 
-from ...utils.utils import get_columns_for_level
+from flowmachine.utils import get_columns_for_level
 
 
 class _TotalCellEvents(Query):
