+++ flowmachine/flowmachine/features/subscriber/day_trajectories.py
@@ -16,7 +16,7 @@ from typing import List
 
 from flowmachine.core import Query
 from flowmachine.features.utilities.subscriber_locations import BaseLocation
-from flowmachine.utils.utils import get_columns_for_level
+from flowmachine.utils import get_columns_for_level
 from ..utilities.multilocation import MultiLocation
