+++ flowmachine/flowmachine/features/subscriber/modal_location.py
@@ -16,7 +16,7 @@ from functools import reduce
 
 from flowmachine.core import Query
 from flowmachine.features.utilities.subscriber_locations import BaseLocation
-from flowmachine.utils.utils import get_columns_for_level
+from flowmachine.utils import get_columns_for_level
 from ..utilities.multilocation import MultiLocation
