+++ flowmachine/flowmachine/features/subscriber/subscriber_call_durations.py
@@ -13,7 +13,7 @@ import warnings
 from typing import List
 
 from ...core import JoinToLocation
-from ...utils.utils import get_columns_for_level
+from flowmachine.utils import get_columns_for_level
 from ..utilities import EventsTablesUnion
 from .metaclasses import SubscriberFeature
