+++ flowmachine/flowmachine/utils.py
@@ -11,6 +11,8 @@ import datetime
 import logging
 from contextlib import contextmanager
 from pathlib import Path
+from pglast import prettify
+from psycopg2._psycopg import adapt
 from threading import get_ident
 from typing import List, Union
 
@@ -18,8 +20,8 @@ import redis_lock
 from redis_lock import AlreadyAcquired
 
 import flowmachine
-from ..core.errors import BadLevelError
-from ..core.errors.flowmachine_errors import MissingColumnsError
+from flowmachine.core.errors import BadLevelError
+from flowmachine.core.errors.flowmachine_errors import MissingColumnsError
 
 logger = logging.getLogger("flowmachine").getChild(__name__)
 
@@ -321,3 +323,58 @@ def rlock(redis_client, lock_id, holder_id=None):
     except AlreadyAcquired:
         yield
     logger.debug(f"{holder_id} released lock id {lock_id}.")
+
+
+def pretty_sql(
+    sql,
+    compact_lists_margin=0,
+    split_string_literals_threshold=0,
+    special_functions=True,
+    comma_at_eoln=True,
+):
+    """
+    Prettify and validate the syntax of an SQL query, using pglast
+
+    Parameters
+    ----------
+    sql : str
+        SQL to prettyify and validate
+    compact_lists_margin : int, default 0
+        Use compact form for lists shorter than this
+    split_string_literals_threshold : int, default 0
+        Split strings (in the sql) longer than this threshold
+    special_functions : bool, default True
+        Translate some special functions to their more commonly used forms
+    comma_at_eoln
+
+    Raises
+    ------
+
+    pglast.parser.ParseError
+        Raises a parse error if the query syntax was bad.
+
+    See Also
+    --------
+    pglast.prettify: Function wrapped, use this for additional prettification options
+
+    Returns
+    -------
+    str
+        The prettified string.
+
+    """
+
+    return prettify(
+        sql,
+        compact_lists_margin=compact_lists_margin,
+        split_string_literals_threshold=split_string_literals_threshold,
+        special_functions=special_functions,
+        comma_at_eoln=comma_at_eoln,
+    )
+
+
+def _makesafe(x):
+    """
+    Function that converts input into a PostgreSQL readable.
+    """
+    return adapt(x).getquoted().decode()
