--- flowmachine/flowmachine/utils/__init__.py
@@ -1,11 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-
-"""
-Small utility functions.
-"""
-
-from .utils import list_of_dates, time_period_add
-
-__all__ = ["list_of_dates", "time_period_add"]
