+++ flowmachine/tests/functional_tests/test_sql_strings_and_results.py
@@ -2,7 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from flowmachine.core.utils import pretty_sql
+from flowmachine.utils import pretty_sql
 
 from approvaltests.approvals import verify
 from flowmachine.core import CustomQuery
